other/robotools/requirements.txt
numpy==1.21.2
Pillow==8.3.1
scipy==1.7.1
trimesh==3.9.20
